changing print flag to comply with holberton checker

+++ 0x1D-search_algorithms/102-interpolation.c
@@ -27,11 +27,11 @@ int interpolation_search(int *array, size_t size, int value)
 
 		if (pos >= size)
 		{
-			printf("Value checked array[%zu] is out of range\n", pos);
+			printf("Value checked array[%u] is out of range\n", (unsigned int)pos);
 			return (-1);
 		}
 
-		printf("Value checked array[%zu] = [%i]\n", pos, array[pos]);
+		printf("Value checked array[%u] = [%i]\n", (unsigned int)pos, array[pos]);
 
 		if (array[pos] == value)
 		{
